robot client refactor
